# Auction/auctionSwagger.yaml
openapi: '3.0.0'
info:
  version: 3.0.0
  title: Auction API's
  description: A sample API
  contact:
      name: API Support
      email: sunil.m@yash.com
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://{base_url}/{env_id}
    variables:
      base_url:
        default: localhost:8080
        description: server
      env_id:
        default: api
        description: path selector
paths:
...
  /auctions/{auction_id}/register:
    post:
        tags:
        - Auction Team Management 
        summary: request to add team to auction
        description: To add New Team in Auction.
        operationId: requestToRegister
        parameters:
        - in: path
          name: auction_id
          required: true
          schema:
            type: integer
        requestBody:
          description: Model to post auction team details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuctionTeamRequestBody'     
        responses:
          200:
            description: "Auction Team has requested to register"
            headers:
              Access-Control-Allow-Origin:
                schema:
                  type: "string"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/AuctionTeamResponseBody"
          204:
            description: "No Content"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/NoContent"        
          400:
            description: "Bad Request"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/BadRequest"
          401:
            description: "Unauthorized"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Unauthorized"
          403:
            description: "Forbidden"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Forbidden"
          404:
            description: "Page Not Found"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/NotFound"        
          500:
            description: "Internal Server Error"
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/InternalServerError"
...
components:
  schemas:
...
    InternalServerError:
      title: Internal server error occurred
      type: object
      properties:
        message:
          type: string
          example: Internal Server Error      
    Unauthorized:
      title: Unauthorized
      type: object
      properties:
        message:
          type: string
          example: Unauthorized
    Forbidden:
      title: Forbidden
      type: object
      properties:
        message:
          type: string
          example: Forbidden
    NoContent:
      title: NoContent
      type: object
      properties:
        message:
          type: string
          example: No Content
    BadRequest:
      title: Invalid input parameters
      type: object
      properties:
        message:
          type: string
          example: Bad Request
...
    AuctionTeamRequestBody:
      type: object
      properties:
        at_team_id:
          type: integer
          example: 1       
    AuctionTeamResponseBody:
      type: object
      properties:
        at_auction_id:
          type: integer
          example: 1
        at_team_id:
          type: integer
          example: 201
        at_approval_status:
          type: string
          example: "pending"
...
    NotFound:
      title: NotFound
      type: object
      properties:
        message:
          type: string
          example: Not Found
...
  securitySchemes:
    AuctionAuthorizer:
      type: "apiKey"
      name: "Authorization"
      in: "header"
      x-amazon-apigateway-authtype: "custom"
      x-amazon-apigateway-authorizer:        
        todo-authorizerUri: 
          Fn::Sub: 'todo-auth'
        authorizerResultTtlInSeconds: 0
        identitySource: "method.request.header.Authorization"
        type: "request"